Fix remaining type mismatch compilation errors
- Fix LocalDateTime parameter type in CrawlerService.updateJobStatus
- Add null check for JobResponse.jobId in JobPostingIntegrationService
- Convert Double to Float for salary in JobRequest mapping
- Convert LocalDate to String for deadline in JobRequest mapping
- Convert Int to Long for totalElements in AdvancedJobSearchService

+++ backend/src/main/kotlin/com/yourjob/backend/service/jobs/AdvancedJobSearchService.kt
@@ -51,14 +51,14 @@ class AdvancedJobSearchService(
             
             // 검색 통계 정보 (첫 페이지일 때만)
             val searchStats = if (filter.page == 0) {
-                generateSearchStats(filter, totalCount)
+                generateSearchStats(filter, totalCount.toLong())
             } else null
             
             return JobSearchResult(
                 content = jobItems,
                 page = filter.page,
                 size = filter.size,
-                totalElements = totalCount,
+                totalElements = totalCount.toLong(),
                 totalPages = totalPages,
                 first = filter.page == 0,
                 last = filter.page >= totalPages - 1,
